+++ man-roxygen/param_ids.R
@@ -1,0 +1,3 @@
+#' @param ids `(character())` \cr
+#' Optional ids for wrapped distributions in vector, should be unique and of same length as
+#' the number of distributions.